Deferring the handling of building arguments, and how to read those arguments, to each integrator
