+++ firecrown/models/cluster/mass_proxy/gaussian.py
@@ -1,9 +1,7 @@
 from typing import List
-
 import numpy as np
 from scipy import special
-
-from firecrown.models.cluster.kernel import Kernel, KernelType, ArgsMapping
+from firecrown.models.cluster.kernel import ArgReader, Kernel, KernelType
 
 
 class MassRichnessGaussian(Kernel):
@@ -15,7 +13,7 @@ class MassRichnessGaussian(Kernel):
         """Return observed scatter corrected by redshift and mass."""
         return NotImplementedError
 
-    def _distribution_binned(self, args: List[float], args_map: ArgsMapping):
+    def _distribution_binned(self, args: List[float], args_map: ArgReader):
         mass = args_map.get_integral_bounds(args, KernelType.mass)
         z = args_map.get_integral_bounds(args, KernelType.z)
         mass_proxy_limits = args_map.get_extra_args(args, self.kernel_type)
@@ -43,7 +41,7 @@ class MassRichnessGaussian(Kernel):
 
         return return_vals
 
-    def _distribution_unbinned(self, args: List[float], args_map: ArgsMapping):
+    def _distribution_unbinned(self, args: List[float], args_map: ArgReader):
         mass = args_map.get_integral_bounds(args, KernelType.mass)
         z = args_map.get_integral_bounds(args, KernelType.z)
         mass_proxy = args_map.get_extra_args(args, self.kernel_type)
